requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
